requirements.txt
DateTime==4.7
dill==0.3.6
gitdb==4.0.9
GitPython==3.1.29
multiprocess==0.70.14
mysqlclient==2.1.1
PyMySQL==1.0.2
pytz==2022.6
smmap==5.0.0
sqlescapy==1.0.1
zope.interface==5.5.1
